+++ cmd/entity/user.go
@@ -5,7 +5,7 @@ import (
 )
 
 type User struct {
-	ID        int32 `gorm:"primaryKey"`
+	ID        int64 `gorm:"primaryKey"`
 	CreatedAt time.Time
 	Uuid      string    `gorm:"uniqueIndex"`
 	Articles  []Article `gorm:"foreignKey:UserId;references:Uuid;constraint:OnUpdate:CASCADE,OnDelete:SET NULL;"`
